refactor(timeline): add useTimelineEvents hook and clean up timeline event logs
- Introduced useTimelineEvents hook for managing project timeline events with persistence and real-time updates.
- Removed numerous console.log debug statements related to timeline events and project actions.
- Updated timeline event storage and event emitter to reduce console noise.
- Improved event handling and state management in timeline-related components.
- Enhanced separation of system and custom timeline events for better event tracking.
